Fix View layer
- Componentizes index to each block of lists
- Creates View Controllers to deal with each content composition
- Upgrades JQuery library
- Renames to View and Rest Controllers and organizes packages
- Loads components through JQuery
