+++ src/main/resources/templates/awards.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<body>
+<div th:fragment="content">
+	<div class="mt-5">
+	    <p>Total Dundie Awards: <span id="totalDundieAwards" th:text="${totalDundieAwards}"></span></p>
+	</div>
+</div>
+</body>
+</html>
