+++ src/main/resources/templates/index.html
@@ -14,79 +14,14 @@
     </style>
 </head>
 <body>
-<div class="mt-5">
-    <p>Total Dundie Awards: <span th:text="${totalDundieAwards}"></span></p>
-</div>
+<div id="awards"></div>
+<div id="employees"></div>
+<div id="activities"></div>
+<div id="queueMessages"></div>
 
-<div th:switch="${employees}" class="mt-5">
-    <p th:case="null" class="alert alert-warning">No employees</p>
-    <div th:case="*">
-        <h2 class="mb-3">Employees</h2>
-        <table class="table table-striped">
-            <thead class="thead-dark">
-            <tr>
-                <th>First name</th>
-                <th>Last name</th>
-                <th>Dundie Awards</th>
-                <th>Organization id</th>
-                <th>Organization name</th>
-            </tr>
-            </thead>
-            <tbody>
-            <tr th:each="employee : ${employees}">
-                <td th:text="${employee.firstName}"></td>
-                <td th:text="${employee.lastName}"></td>
-                <td th:text="${employee.dundieAwards}"></td>
-                <td th:text="${employee.organizationId}"></td>
-                <td th:text="${employee.organizationName}"></td>
-            </tr>
-            </tbody>
-        </table>
-    </div>
-</div>
-<div th:switch="${activities}" class="mt-5">
-    <p th:case="null" class="alert alert-warning">No activities</p>
-    <div th:case="*">
-        <h2 class="mb-3">Activity Log</h2>
-        <table class="table table-striped">
-            <thead class="thead-dark">
-            <tr>
-                <th>Occurred At</th>
-                <th>Event</th>
-            </tr>
-            </thead>
-            <tbody>
-            <tr th:each="activity : ${activities}">
-                <th th:text="${activity.occurredAt}"></th>
-                <th th:text="${activity.event}"></th>
-            </tr>
-            </tbody>
-        </table>
-    </div>
-</div>
-<div th:switch="${queueMessages}" class="mt-5">
-    <p th:case="null" class="alert alert-warning">No messages</p>
-    <div th:case="*">
-        <h2 class="mb-3">Queue Messages</h2>
-        <table class="table table-striped">
-            <thead class="thead-dark">
-            <tr>
-                <th>Occurred At</th>
-                <th>Event</th>
-            </tr>
-            </thead>
-            <tbody>
-            <tr th:each="activity : ${queueMessages}">
-                <th th:text="${activity.occurredAt}"></th>
-                <th th:text="${activity.event}"></th>
-            </tr>
-            </tbody>
-        </table>
-    </div>
-</div>
 <!-- Include Bootstrap JS and Popper.js -->
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
-        integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+<script src="https://code.jquery.com/jquery-3.7.1.min.js"
+        integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo="
         crossorigin="anonymous"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
         integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
@@ -95,5 +30,53 @@
         integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
         crossorigin="anonymous"></script>
 <script th:if="${isDev}" src="http://localhost:35729/livereload.js"></script>
+<script>
+
+	function getTotalDundieAwards() {
+		$.get({
+            url: '/awards/view',
+            success: (response) => {
+            }
+        }).done((data) => {
+            $('#awards').html(data);
+        });
+	}
+	
+     function getEmployees() {
+     	$.get({
+             url: '/employees/view'
+         }).done((data) => {
+             $('#employees').html(data);
+         });
+     };
+     
+      function getActivities() {
+          $.get({
+              url: '/activities/view',
+              success: (response) => {
+              }
+          }).done((data) => {
+              $('#activities').html(data);
+          });
+      };
+      
+      function getMessages() {
+          $.get({
+              url: '/messages/view',
+              success: (response) => {
+              }
+          }).done((data) => {
+              $('#queueMessages').html(data);
+          });
+      };
+
+     $(document).ready(() => {
+    	 getTotalDundieAwards();
+    	 getEmployees();
+    	 getActivities();
+    	 getMessages();
+     });
+     
+</script>
 </body>
 </html>
